Customer places an order with multiple items System validates stock is available for each item Stock is deducted from inventory atomically Order status progresses: PENDING → CONFIRMED → SHIPPED → DELIVERED (or CANCELLED) Cancellation restores stock

+++ src/main/java/com/nebulastock/nebulastock/dto/response/InventoryResponse.java
@@ -10,11 +10,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class InventoryResponse {
-    private Integer id;
-    private Integer productId;
+    private Long id;
+    private Long productId;
     private String productName;
     private String productSku;
-    private Integer warehouseId;
+    private Long warehouseId;
     private String warehouseName;
     private Integer quantity;
 }
