+++ src/main/java/com/nebulastock/nebulastock/repository/OrderItemRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
-    List<OrderItem> findByOrderId(Integer orderId);
+    List<OrderItem> findByOrderId(Long orderId);
 }
